Finish migrating tests from mixture module to seq module
