+++ README.md
@@ -56,8 +56,8 @@ Invoke `mhpl8r <subcmd> --help` and replace `<subcmd>` with one of the
 subcommands listed below to see instructions for that operation.
 
 Subcommands:
-  subcmd             contain, contrib, dist, getrefr, mixture, prob, refr,
-                     seq, sim, type
+  subcmd             contain, contrib, dist, getrefr, prob, refr, seq, sim,
+                     type
 
 Global arguments:
   -h, --help         show this help message and exit
