+++ microhapulator/panel.py
@@ -229,9 +229,10 @@ def validate_populations(popids):
     if len(popids) not in (1, 2):
         message = 'please provide only 1 or 2 population IDs'
         raise ValueError(message)
+    popids = sorted(set(popids))
     haplopops = microhapdb.standardize_ids(popids)
-    if not haplopops <= popids:
-        raise ValueError('invalid population ID(s)')
+    if len(haplopops) < len(popids):
+        raise ValueError('invalid or duplicated population ID(s)')
     if len(haplopops) == 1:
         haplopops = haplopops * 2
     return haplopops
